+++ src/functions/arithmetic/multiply.js
@@ -12,6 +12,7 @@ class LimeFunctionMultiply extends LimeFunction {
 
     // Binary operation
     this.operations.b = [
+      'tr(_!)',
       'tr(_+)',
       'tr(_-)',