+++ src/functions/arithmetic/negative.js
@@ -12,6 +12,7 @@ class LimeFunctionNegative extends LimeFunction {
 
     // Right unary operation
     this.operations.r = [
+      'tr(_!)',
       'tr(_+)',
       'tr(_-)',