+++ src/functions/bitwise/shiftLeft.js
@@ -3,36 +3,28 @@ const LimeFunction = require('../../structs/function');
 
 /* ------------------------ division ------------------------ */
 
-// Greater function class
-class LimeFunctionGreater extends LimeFunction {
+// Shift left function class
+class LimeFunctionShiftLeft extends LimeFunction {
   // Constructor
   constructor(lime, mode) {
     // Super from function class
-    super(lime, { name: 'greater', mode });
+    super(lime, { name: 'shiftLeft', mode });
 
     // Binary operation
     this.operations.b = [
-      'tf(>=)',
-      'tf(>>)',
       'tr(_!)',
       'tr(_+)',
       'tr(_-)',
 
-      'cb(int->rat,rat)',
-      'cb(rat,int->rat)',
       'cl(var->exp)',
       'cr(var->exp)',
 
       'eb(int,int)',
-      'eb(rat,rat)',
     ];
 
     // Algorithms
     this.algorithms.set('b(int,int)', (step) => {
-      step.bs(this.lime.build('boolean')(step.left.value > step.right.value));
-    });
-    this.algorithms.set('b(rat,rat)', (step) => {
-      step.bs(this.lime.direct([step.left.nPlace, '*', step.right.dPlace, '>', step.right.nPlace, '*', step.left.dPlace]));
+      step.bs(this.lime.direct([step.left.value << step.right.value]));
     });
   }
 }
@@ -40,4 +32,4 @@ class LimeFunctionGreater extends LimeFunction {
 /* ------------------------ division ------------------------ */
 
 // Export
-module.exports = LimeFunctionGreater;
+module.exports = LimeFunctionShiftLeft;