+++ src/functions/comparison/equal.js
@@ -12,6 +12,7 @@ class LimeFunctionEqual extends LimeFunction {
 
     // Binary operation
     this.operations.b = [
+      'tr(_!)',
       'tr(_+)',
       'tr(_-)',