+++ src/functions/comparison/notEqual.js
@@ -13,6 +13,7 @@ class LimeFunctionNotEqual extends LimeFunction {
     // Binary operation
     this.operations.b = [
       'tf(!==)',
+      'tr(_!)',
       'tr(_+)',
       'tr(_-)',