+++ src/functions/comparison/smallerEqual.js
@@ -12,6 +12,7 @@ class LimeFunctionSmallerEqual extends LimeFunction {
 
     // Binary operation
     this.operations.b = [
+      'tr(_!)',
       'tr(_+)',
       'tr(_-)',